src/pipeline/preprocess.py
from PIL import Image
import numpy as np

# Functions for preprocessing images before they enter the AI pipeline.

def preprocess(args):
    """
    Placeholder function to preprocess an image.
    It handles both a file path from the CLI (args.input) and a
    NumPy array from the Gradio GUI (args.input_image).
    """
    if hasattr(args, 'input_image') and args.input_image is not None:
        print("Preprocessing image from Gradio input...")
        # The input is already a NumPy array, so we can use it directly.
        img = args.input_image
        # In a real scenario, you might convert it to a PIL Image or other format
        # img = Image.fromarray(img) 
    elif hasattr(args, 'input') and args.input:
        print(f"Preprocessing image from file: {args.input}")
        # Here you would load the image from the file path
        # For this placeholder, we'll create a dummy numpy array.
        try:
            img = np.array(Image.open(args.input))
        except FileNotFoundError:
            print(f"Error: File not found at {args.input}")
            return None
    else:
        print("Error: No valid input provided.")
        return None

    # Returns a dummy numpy array for now, but in reality, it would be the preprocessed image
    print("Image preprocessed successfully.")
    return img