requirements.txt
# Placeholder for requirements
fastapi
uvicorn
gfpgan
basicsr
numpy
pillow
torch
torchvision
torchaudio
gradio
pyinstaller
